backend/pyproject.toml
[tool.poetry]
name = "report-creator"
version = "0.1.0"
description = ""
authors = ["Ryo Nagaoka"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.112.0"
uvicorn = "0.30.5"
pydantic = "2.8.2"
python-dotenv = "1.0.1"
google-generativeai = "0.7.2"
tiktoken = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
